pull request peer review changes(#476)
- rename NetconfClient to NetconfSSHClient
- add const declaration
- fix simple tcp proxy server recv method
