+++ sdk/cpp/core/src/netconf_provider.cpp
@@ -29,7 +29,7 @@
 #include "entity_data_node_walker.hpp"
 #include "errors.hpp"
 #include "ietf_parser.hpp"
-#include "netconf_client.hpp"
+#include "netconf_ssh_client.hpp"
 #include "netconf_tcp_client.hpp"
 #include "netconf_provider.hpp"
 #include "netconf_model_provider.hpp"
@@ -53,7 +53,7 @@ static void create_input_source(path::DataNode & input, bool config);
 static void create_input_error_option(path::DataNode & input);
 static string get_annotated_config_payload(path::RootSchemaNode & root_schema, path::Rpc & rpc, path::Annotation & annotation);
 static string get_commit_rpc_payload();
-static std::shared_ptr<path::DataNode> handle_edit_reply(string reply, NetconfClientBase & client, bool candidate_supported);
+static std::shared_ptr<path::DataNode> handle_edit_reply(string reply, NetconfClient & client, bool candidate_supported);
 
 static string get_read_rpc_name(bool config);
 static bool is_config(path::Rpc & rpc);
@@ -65,7 +65,7 @@ const char* CANDIDATE = "urn:ietf:params:netconf:capability:candidate:1.0";
 const string PROTOCOL_SSH = "ssh";
 const string PROTOCOL_TCP = "tcp";
 
-NetconfServiceProvider::NetconfServiceProvider(string address, string username, string password, int port, std::string protocol)
+NetconfServiceProvider::NetconfServiceProvider(const string& address, const string& username, const string& password, int port, const string& protocol)
 {
     initialize_client(address, username, password, port, protocol);
     path::Repository repo;
@@ -73,18 +73,18 @@ NetconfServiceProvider::NetconfServiceProvider(string address, string username, 
     YLOG_INFO("Connected to {} on port {} using {}", address, port, protocol);
 }
 
-NetconfServiceProvider::NetconfServiceProvider(path::Repository & repo, string address, string username, string password, int port, std::string protocol)
+NetconfServiceProvider::NetconfServiceProvider(path::Repository & repo, const string& address, const string& username, const string& password, int port, const string& protocol)
 {
     initialize_client(address, username, password, port, protocol);
     initialize(repo);
     YLOG_INFO("Connected to {} on port {} using {}", address, port, protocol);
 }
 
-void NetconfServiceProvider::initialize_client(string address, string username, string password, int port, std::string protocol)
+void NetconfServiceProvider::initialize_client(const string& address, const string& username, const string& password, int port, const string& protocol)
 {
     if (protocol.compare(PROTOCOL_SSH) == 0)
     {
-        client = make_unique<NetconfClient>(username, password, address, port);
+        client = make_unique<NetconfSSHClient>(username, password, address, port);
     }
     else if (protocol.compare(PROTOCOL_TCP) == 0)
     {
@@ -349,7 +349,7 @@ static string get_netconf_payload(path::DataNode & input, string data_tag, strin
     return payload;
 }
 
-static std::shared_ptr<path::DataNode> handle_edit_reply(string reply, NetconfClientBase & client, bool candidate_supported)
+static std::shared_ptr<path::DataNode> handle_edit_reply(string reply, NetconfClient & client, bool candidate_supported)
 {
     if(reply.find("<ok/>") == std::string::npos)
     {
