+++ sdk/cpp/tests/CMakeLists.txt
@@ -8,7 +8,7 @@ enable_testing()
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${CMAKE_SOURCE_DIR}/CMakeModules/")
 set(bundle_tests_src
         test_core_netconf.cpp
-        test_netconf_client.cpp
+        test_netconf_ssh_client.cpp
         test_netconf_tcp_client.cpp
         test_restconf_client.cpp
         testsanitynctest.cpp
